Implemented a hbbft validator parity toml config generator
