+++ ethcore/hbbft_engine/hbbft_config_generator/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+description = "parity config generator for hbbft validators"
+name = "hbbft_config_generator"
+version = "0.0.1"
+license = "GPL-3.0"
+authors = [
+    "David Forstenlechner <dforsten@gmail.com>"
+]
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+hbbft = { git = "https://github.com/poanetwork/hbbft" }
+hbbft_testing = { git = "https://github.com/poanetwork/hbbft" }
+toml = "0.5.1"
+rand = "0.6.5"
+clap = "2"
+
+[dev-dependencies]
